+++ extruntime/extruntime_darwin.go
@@ -4,7 +4,12 @@ package extruntime
 
 import (
 	"github.com/rs/zerolog"
+	"github.com/rs/zerolog/log"
 )
+
+func adjustOOMScoreAdj(_ int) {
+	log.Debug().Msg("oom_score_adj adjustment is not supported on this platform")
+}
 
 func logCapsInformation(_ zerolog.Level) {
 }